requirements.txt
requests==2.26.0
Django==3.2
djangorestframework==3.12.4
PyJWT==1.7.1
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
drf-yasg==1.21.5
djangorestframework-simplejwt==5.2.2
django-filter==21.1
